+++ src/contenedores/Notificaciones.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import { obtenerOrdenesFiltradas } from '../firebase/firebaseStore';
-import BarraPrincipal from '../componente/BarraPrincipal';
-import CabeceraUsuario from '../componente/CabeceraUsuario';
-import BarraSecundaria from '../componente/BarraSecundaria';
-import PedidoPendiente from '../componente/PedidoPendiente';
-import '../estilo/Notificaciones.scss';
+import CabeceraUsuario from '../componentes/CabeceraUsuario';
+import BarraSecundaria from '../componentes/BarraSecundaria';
+import OrdenPendiente from '../componentes/OrdenPendiente';
+import '../estilos/Notificaciones.scss';
 
-export default function Pedidos() {
-  const [subcategoria, setSubcategoria] = useState('Por preparar');
+export default function Notificaciones() {
+  // Estado de la subcategoria
+  const [subcategoria, setSubcategoria] = useState('Por entregar');
 
   // Funcion obtener productos de Firestore
   const bdOrdenes = async () => {
@@ -21,19 +21,21 @@ export default function Pedidos() {
 
   // Estado del arreglo de productos
   const [arregloOrdenes, setArregloOrdenes] = useState([]);
+
   // Efecto para que se actualice los productos cuando cambie la subcategoria
   useEffect(() => bdOrdenes().then((arr) => {
     console.log(arr);
     setArregloOrdenes(arr);
   }), [subcategoria]);
+
   return (
-    <div className="Pedidos">
-      <BarraPrincipal vista="Pedidos" />
+    <div className="Notificaciones">
       <CabeceraUsuario />
-      <BarraSecundaria cat="Pedidos" setSubcategoria={setSubcategoria} />
-      <div className="contenedorPedidos">
-        {arregloOrdenes.map((orden) => <PedidoPendiente key={orden.numeroOrden} orden={orden} subcategoria={subcategoria} />)}
+      <BarraSecundaria cat="Notificaciones" setSubcategoria={setSubcategoria} />
+      <div className="contenedorOrdenes">
+        {arregloOrdenes.map((orden) => <OrdenPendiente key={orden.numeroOrden} orden={orden} subcategoria={subcategoria} />)}
       </div>
+
     </div>
   );
 }
