+++ src/estilos/BarraSecundaria.scss
