+++ src/estilos/DetalleOrden.scss
