+++ src/paginas/InicioSesion.jsx
@@ -2,11 +2,11 @@ import React, { useRef, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faUserCircle } from '@fortawesome/free-solid-svg-icons';
-import AppContext from '../contexto/AppContext';
+import AppContext from '../contextos/AppContext';
 import { iniciarSesion } from '../firebase/firebaseAuth';
 import { obtenerUsuario } from '../firebase/firebaseStore';
-import logo from '../imagen/logoHamburguesa.svg';
-import '../estilo/InicioSesion.scss';
+import logo from '../imagenes/logoHamburguesa.svg';
+import '../estilos/InicioSesion.scss';
 
 export default function InicioSesion() {
   const { obtenerNombreMesero } = useContext(AppContext);
